Refactor: Message input validations (#478)
* Refactor: Remove dependency on entire message

* Move connection type validation from master data document to commands

* Refactor: disconnection type validation rule

* Move disconnection type validation from master data document to commands

* Move gsrn number validation from master data document to commands

* Refactor settlement method validation

* Move test

* Remove settlement method validation from master data document to command

* Refactor metering method validation rule

* Refactor meter number validation rule

* Refactor net settlement group validation rule

* Refactor product type validation rule

* Refactor validation rule for capacity

* Refactor location description validation rule

* Refactor power plant validation rule

* Refactor asset type validation rule

* Move country code validation rule to command level

* Remove duplicate rule definition

* Move effective date validation rule definition to command level

* Remove test since the rule is enforced at domain level

* Fix broken test

* Fix broken test

* Fix broken test

* Fix broken test

* Fix broken test

* Move test

* Move test

* Move test

* Move test

* Move test

* Move test

* Remove obsolete test class

* Remove obsolete class

* Remove obsolete class

* Remove obsolete objects

* Remove unnecessary cascade definition
