+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/Errors/Converters/ProductTypeInvalidValueErrorConverter.cs
@@ -24,7 +24,7 @@ namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.Errors.Converters
         {
             if (validationError == null) throw new ArgumentNullException(nameof(validationError));
 
-            return new ErrorMessage("D02", $"Product {validationError.ProductType} for metering point {validationError.GsrnNumber} has wrong value (outside domain)");
+            return new ErrorMessage("D02", $"Product {validationError.ProductType} has wrong value (outside domain)");
         }
     }
 }
