--- source/Energinet.DataHub.MeteringPoints.IntegrationTests/MarketDocuments/TestBusinessRequest.cs
@@ -1,35 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System;
-using Energinet.DataHub.MeteringPoints.Application.Common;
-
-namespace Energinet.DataHub.MeteringPoints.IntegrationTests.MarketDocuments
-{
-    public class TestBusinessRequest : IBusinessRequest
-    {
-        public TestBusinessRequest(string transactionId, string effectiveDate, string gsrnNumber)
-        {
-            TransactionId = transactionId;
-            EffectiveDate = effectiveDate;
-            GsrnNumber = gsrnNumber;
-        }
-
-        public string TransactionId { get; }
-
-        public string EffectiveDate { get; }
-
-        public string GsrnNumber { get; }
-    }
-}
